requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
openai>=1.14.0
pymilvus>=2.3.6
pandas>=2.2.0
openpyxl>=3.1.2
tenacity>=8.2.3
redis>=5.0.1
structlog>=24.1.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
pybreaker>=1.2.0
langsmith>=0.1.0
openevals>=0.0.1
langchain-openai>=0.1.0
